pyproject.toml
[project]
name = "cerebro"
version = "0.1.0"
description = "Personal AI assistant built on LangGraph, powered by local-brain."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "langgraph>=1.0.0",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langchain>=0.2.14",
    "langchain-anthropic>=0.1.23",
    "langchain-tavily>=0.1",
    "langchain-mcp-adapters>=0.1.0",
    "textual>=0.80.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
cerebro = "cerebro.chat:main"

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["cerebro"]

[tool.setuptools.package-dir]
"cerebro" = "src/cerebro"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.10",
    "pytest>=8.3.5",
]
